Update MetaTags.js
applied meta language tag to fix auto-translate popups.

As AntD doesnt expose the html tag, modern browsers fall back to the meta tag is none is set:

https://www.w3.org/International/questions/qa-http-and-lang#meta

Until recently, few browsers paid any attention to this meta element. Then several major browsers began using this element, if there was no language attribute on the html tag, to set the default language of the text in the document (what you should use a language attribute on the html tag for). The way this was implemented was inconsistent, and therefore unreliable, from one browser to another.

HTML5 did, however, make a concession for backwards compatibility. If there is a meta element with http-equiv set to Content-Language in the markup, and if there is no language attribute on the html tag, and if the meta element has a value that is a single language tag, then a browser can (not must) use that information to guess at the default language of the text on the page. Having said that, this is only for backwards compatibility, and you really shouldn't use this approach any more. Simply use a language attribute on the html tag.